# spring/api-docs.yaml
openapi: 3.0.1
info:
  title: Конспект по языку Java
  description: Конспект по языку Java на русском языке
  contact:
    name: Токовенко Виктор
    url: https://github.com/Lorgar-Avrelian?tab=repositories
    email: victor-14-244@mail.ru
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /readers:
    get:
      tags:
        - Поиск
        - 3 Читатели
      summary: Список читателей
      description: Вывести список всех зарегистрированных читателей
      operationId: getAllReaders
      parameters:
        - name: partOfNameSecondNameOrSurname
          in: query
          description: Часть ФИО читателя
          required: false
          schema:
            type: string
          example: Иван
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReaderNoBooksDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
    put:
      tags:
        - 3 Читатели
        - Читатель
      summary: Редактировать
      description: Отредактировать информацию о читателе
      operationId: updateReader
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reader"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReaderNoBooksDTO"
        "400":
          description: Bad Request
      security:
        - basicAuth: []
    post:
      tags:
        - 3 Читатели
        - Читатель
      summary: Создать
      description: Добавить информацию о читателе
      operationId: createReader
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewReaderDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReaderNoBooksDTO"
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      security:
        - basicAuth: []
  /books:
    get:
      tags:
        - 2 Книги
        - Поиск
      summary: Список книг
      description: Вывести список всех доступных книг
      operationId: getAllBooks
      parameters:
        - name: authorOrTitle
          in: query
          description: Часть ФИО автора или названия книги
          required: false
          schema:
            type: string
          example: Пушкин
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
    put:
      tags:
        - 2 Книги
        - Книга
      summary: Редактировать
      description: Отредактировать информацию о книге
      operationId: updateBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDTO"
        "400":
          description: Bad Request
      security:
        - basicAuth: []
    post:
      tags:
        - 2 Книги
        - Книга
      summary: Создать
      description: Добавить информацию о книге
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBookDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDTO"
        "405":
          description: Method Not Allowed
      security:
        - basicAuth: []
...
  /readers/{id}:
    get:
      tags:
        - 3 Читатели
        - Читатель
      summary: Найти
      description: Найти информацию о читателе по ID
      operationId: readReader
      parameters:
        - name: id
          in: path
          description: ID читателя в имеющемся списке читателей
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReaderNoBooksDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
    delete:
      tags:
        - 3 Читатели
        - Читатель
      summary: Удалить
      description: Удалить информацию о читателе по ID
      operationId: deleteReader
      parameters:
        - name: id
          in: path
          description: ID читателя в имеющемся списке читателей
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReaderNoBooksDTO"
        "403":
          description: Forbidden
      security:
        - basicAuth: []
  /readers/surname:
    get:
      tags:
        - Поиск
        - 3 Читатели
      summary: Список по фамилии
      description: Вывести список всех зарегистрированных читателей с указанной фамилией
      operationId: getReadersBySurname
      parameters:
        - name: partOfSurname
          in: query
          description: Часть фамилии читателя
          required: true
          schema:
            type: string
          example: Иванов
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReaderNoBooksDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
  /readers/secondName:
    get:
      tags:
        - Поиск
        - 3 Читатели
      summary: Список по отчеству
      description: Вывести список всех зарегистрированных читателей с указанным отчеством
      operationId: getReadersBySecondName
      parameters:
        - name: partOfSecondName
          in: query
          description: Часть отчества читателя
          required: true
          schema:
            type: string
          example: Иванович
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReaderNoBooksDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
  /readers/name:
    get:
      tags:
        - Поиск
        - 3 Читатели
      summary: Список по имени
      description: Вывести список всех зарегистрированных читателей с указанным именем
      operationId: getReadersByName
      parameters:
        - name: partOfName
          in: query
          description: Часть имени читателя
          required: true
          schema:
            type: string
          example: Иван
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReaderNoBooksDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
  /readers/books:
    get:
      tags:
        - Поиск
        - 3 Читатели
      summary: Список книг
      description: Вывести список всех книг данного читателя
      operationId: getReaderBooks
      parameters:
        - name: id
          in: query
          description: ID читателя в имеющемся списке читателей
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
...
  /books/{id}:
    get:
      tags:
        - 2 Книги
        - Книга
      summary: Найти
      description: Найти информацию о книге по ID
      operationId: readBook
      parameters:
        - name: id
          in: path
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
    delete:
      tags:
        - 2 Книги
        - Книга
      summary: Удалить
      description: Удалить информацию о книге по ID
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDTO"
        "403":
          description: Forbidden
      security:
        - basicAuth: []
...
components:
  schemas:
    ReaderNoBooksDTO:
      title: Читатель
      maximum: 9223372036854775807
      minimum: 1
      required:
        - id
        - name
        - personalNumber
        - surname
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID читателя
          format: int64
          default: 1
        name:
          title: Имя
          maxLength: 12
          minLength: 2
          type: string
          description: Имя читателя
          default: Иван
        secondName:
          title: Отчество
          maxLength: 16
          minLength: 2
          type: string
          description: Отчество читателя
          default: Иванович
        surname:
          title: Фамилия
          maxLength: 30
          minLength: 2
          type: string
          description: Фамилия читателя
          default: Иванов
        personalNumber:
          title: Номер билета
          maximum: 2147483648
          minimum: 1
          type: integer
          description: Номер читательского билета
          format: int32
          default: 1
      description: Сущность читателя
      default: null
    Reader:
      title: Читатель
      maximum: 9223372036854775807
      minimum: 1
      required:
        - id
        - name
        - personalNumber
        - surname
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID читателя
          format: int64
          default: 1
        name:
          title: Имя
          maxLength: 12
          minLength: 2
          type: string
          description: Имя читателя
          default: Иван
        secondName:
          title: Отчество
          maxLength: 16
          minLength: 2
          type: string
          description: Отчество читателя
          default: Иванович
        surname:
          title: Фамилия
          maxLength: 30
          minLength: 2
          type: string
          description: Фамилия читателя
          default: Иванов
        personalNumber:
          title: Номер билета
          maximum: 2147483648
          minimum: 1
          type: integer
          description: Номер читательского билета
          format: int32
          default: 1
      description: Сущность читателя
      default: null
    BookDTO:
      title: Книга
      required:
        - author
        - id
        - title
        - year
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID книги
          format: int64
          default: 1
        title:
          title: Название
          maxLength: 30
          minLength: 3
          type: string
          description: Название книги
          default: Война и мир
        author:
          title: Автор
          maxLength: 24
          minLength: 8
          type: string
          description: Автор книги
          default: Л.Н. Толстой
        year:
          title: Год
          minimum: 1970
          type: integer
          description: Год публикации книги
          format: int32
          default: 1986
        reader:
          $ref: "#/components/schemas/ReaderNoBooksDTO"
      description: Сущность книги
...
    NewReaderDTO:
      title: Новый читатель
      required:
        - name
        - personalNumber
        - surname
      type: object
      properties:
        name:
          title: Имя
          maxLength: 12
          minLength: 2
          type: string
          description: Имя читателя
          default: Иван
        secondName:
          title: Отчество
          maxLength: 16
          minLength: 2
          type: string
          description: Отчество читателя
          default: Иванович
        surname:
          title: Фамилия
          maxLength: 30
          minLength: 2
          type: string
          description: Фамилия читателя
          default: Иванов
        personalNumber:
          title: Номер билета
          maximum: 2147483648
          minimum: 1
          type: integer
          description: Номер читательского билета
          format: int32
          default: 1
      description: Сущность нового читателя
...
    NewBookDTO:
      title: Новая книга
      required:
        - authorFIO
        - title
        - year
      type: object
      properties:
        title:
          title: Название
          maxLength: 30
          minLength: 3
          type: string
          description: Название книги
          default: Война и мир
        authorFIO:
          title: Автор
          maxLength: 24
          minLength: 8
          type: string
          description: Автор книги
          default: Л.Н. Толстой
        year:
          title: Год
          minimum: 1970
          type: integer
          description: Год публикации книги
          format: int32
          default: 1986
      description: Сущность новой книги
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic